chore: add node_data/leaf
