+++ crates/committer/src/patricia_merkle_tree/filled_tree/node.rs
@@ -4,7 +4,7 @@ use crate::patricia_merkle_tree::filled_tree::node_serde::{
     LeafCompiledClassToSerialize, SerializeNode,
 };
 use crate::patricia_merkle_tree::node_data::inner_node::NodeData;
-use crate::patricia_merkle_tree::types::LeafDataTrait;
+use crate::patricia_merkle_tree::node_data::leaf::{LeafData, LeafDataTrait};
 use crate::types::Felt;
 
 // TODO(Nimrod, 1/6/2024): Swap to starknet-types-core types once implemented.
@@ -25,36 +25,6 @@ pub(crate) struct CompiledClassHash(pub Felt);
 pub(crate) struct FilledNode<L: LeafDataTrait> {
     pub(crate) hash: HashOutput,
     pub(crate) data: NodeData<L>,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) enum LeafData {
-    StorageValue(Felt),
-    CompiledClassHash(ClassHash),
-    StateTreeTuple {
-        class_hash: ClassHash,
-        contract_state_root_hash: Felt,
-        nonce: Nonce,
-    },
-}
-
-impl LeafDataTrait for LeafData {
-    fn is_empty(&self) -> bool {
-        match self {
-            LeafData::StorageValue(value) => *value == Felt::ZERO,
-            LeafData::CompiledClassHash(class_hash) => class_hash.0 == Felt::ZERO,
-            LeafData::StateTreeTuple {
-                class_hash,
-                contract_state_root_hash,
-                nonce,
-            } => {
-                nonce.0 == Felt::ZERO
-                    && class_hash.0 == Felt::ZERO
-                    && *contract_state_root_hash == Felt::ZERO
-            }
-        }
-    }
 }
 
 impl LeafData {
