+++ crates/committer/src/patricia_merkle_tree/node_data.rs
@@ -1,1 +1,2 @@
 pub mod inner_node;
+pub mod leaf;
