+++ crates/committer/src/patricia_merkle_tree/node_data/leaf.rs
@@ -1,0 +1,37 @@
+use crate::patricia_merkle_tree::filled_tree::node::{ClassHash, Nonce};
+use crate::types::Felt;
+
+pub(crate) trait LeafDataTrait {
+    /// Returns true if leaf is empty.
+    fn is_empty(&self) -> bool;
+}
+
+#[allow(dead_code)]
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub(crate) enum LeafData {
+    StorageValue(Felt),
+    CompiledClassHash(ClassHash),
+    StateTreeTuple {
+        class_hash: ClassHash,
+        contract_state_root_hash: Felt,
+        nonce: Nonce,
+    },
+}
+
+impl LeafDataTrait for LeafData {
+    fn is_empty(&self) -> bool {
+        match self {
+            LeafData::StorageValue(value) => *value == Felt::ZERO,
+            LeafData::CompiledClassHash(class_hash) => class_hash.0 == Felt::ZERO,
+            LeafData::StateTreeTuple {
+                class_hash,
+                contract_state_root_hash,
+                nonce,
+            } => {
+                nonce.0 == Felt::ZERO
+                    && class_hash.0 == Felt::ZERO
+                    && *contract_state_root_hash == Felt::ZERO
+            }
+        }
+    }
+}
