+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/tree.rs
@@ -2,7 +2,8 @@ use std::collections::HashMap;
 
 use crate::hash::hash_trait::HashOutput;
 use crate::patricia_merkle_tree::errors::OriginalSkeletonTreeError;
-use crate::patricia_merkle_tree::types::{LeafDataTrait, NodeIndex, TreeHeight};
+use crate::patricia_merkle_tree::node_data::leaf::LeafDataTrait;
+use crate::patricia_merkle_tree::types::{NodeIndex, TreeHeight};
 use crate::patricia_merkle_tree::updated_skeleton_tree::tree::UpdatedSkeletonTree;
 use crate::storage::storage_trait::Storage;
